some changes

+++ backend/src/main/java/com/matheusmr/tibiaauctionhistory/common/model/AuctionsResult.java
@@ -2,5 +2,5 @@ package com.matheusmr.tibiaauctionhistory.common.model;
 
 import java.util.List;
 
-public record AuctionsResult(List<Auction> auctions, int totalCount) {
+public record AuctionsResult(List<Auction> auctions, long totalCount) {
 }
